Fix Firo Build Warnings (#1872)
* fix: use glibc assert hook with clang on linux

* fix: clean up mechanical warning noise

* fix: clean up stacktrace warning guards

* fix: update spark mint documentation comments

* fix: address compatibility warning cleanup

* fix: initialize warning-prone values

* build: keep project warnings off vendored targets

* build: apply crash hook wraps at link time

* fix: add peer logic validation destructor

* fix: initialize warning-prone byte storage

* fix: remove unreachable rpc post command

* build: mark crypto boost includes as system

* fix: clean up mingw source warnings

* build: scope secp256k1 warning flags to C

* fix: reserve spark ownership proof streams

* fix: clean up automint lock and platform guards

* build: package immer as dependency

* build: report global compile properties as notices

* fix: address remaining build warnings

* build: set macOS deployment target for dependencies

* test: remove flaky MTP nonce uniqueness check

* build: avoid macOS byte-swap macro redefinitions

* Build: add warning interface to bitcoin_crypto

* Tests: fix spark name overflow captures

* Build: fix bitcoin_crypto warning noise

* Build: suppress prevector false positive

+++ src/spark/state.cpp
@@ -335,7 +335,7 @@ bool ConnectBlockSpark(
             try {
                 for (const auto &sparkName : pblock->sparkTxInfo->sparkNames) {
                     uint8_t opType = sparkName.second.nVersion >= 2 ?
-                                                    sparkName.second.operationType : CSparkNameTxData::opRegister;
+                                                    sparkName.second.operationType : static_cast<uint8_t>(CSparkNameTxData::opRegister);
                     // For V2.1+, renewals and transfers preserve remaining validity
                     int validityBlocks = sparkName.second.sparkNameValidityBlocks;
                     const auto& consensusParams = ::Params().GetConsensus();
